Enabling telemetry on nucliadb
