+++ .github/workflows/nucliadb_telemetry.yml
@@ -1,0 +1,81 @@
+name: NucliaDB Telemetry (py)
+
+on:
+  pull_request:
+    branches:
+      - main
+    paths:
+      - ".github/workflows/nucliadb_telemetry.yml"
+      - "nucliadb_telemetry/**"
+  push:
+    branches:
+      - main
+    paths:
+      - ".github/workflows/nucliadb_telemetry.yml"
+      - "nucliadb_telemetry/**"
+
+jobs:
+  # Job to run pre-checks
+  pre-checks:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.9]
+
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v3.1.0
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: "pip"
+
+      - name: Install package
+        run: |
+          pip install -r code-requirements.txt
+      - name: Run pre-checks
+        run: |
+          flake8 nucliadb_telemetry --config=nucliadb_telemetry/setup.cfg
+          isort -c --profile black nucliadb_telemetry/
+          black --check --verbose nucliadb_telemetry
+
+  # Job to run tests
+  tests:
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        python-version: [3.9]
+
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v3.1.0
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: "pip"
+
+      - name: Install the package
+        run: |
+          pip install -r test-requirements.txt
+          pip install -r nucliadb_telemetry/requirements-sources.txt
+          pip install -r nucliadb_telemetry/requirements.txt
+          pip install -e nucliadb_telemetry
+
+      - name: Check mypy
+        run: |
+          MYPYPATH=./mypy_stubs mypy nucliadb_telemetry/
+
+      - name: Run tests
+        run: |
+          pytest -rfE --cov=nucliadb_telemetry -s --tb=native -v --cov-report xml --cov-append nucliadb_telemetry/nucliadb_telemetry
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v3
+        with:
+          file: ./coverage.xml
+          flags: telemetry
