+++ nucliadb_reader/nucliadb_reader/__init__.py
@@ -19,7 +19,9 @@
 #
 import logging
 
-logger = logging.getLogger("nucliadb_reader")
+
+SERVICE_NAME = "nucliadb_reader"
+logger = logging.getLogger(SERVICE_NAME)
 
 API_PREFIX = "api"
