+++ nucliadb_reader/nucliadb_reader/lifecycle.py
@@ -22,12 +22,13 @@ import sys
 
 from nucliadb_ingest.utils import start_ingest, stop_ingest
 from nucliadb_reader import logger
+from nucliadb_reader import SERVICE_NAME
 from nucliadb_utils.settings import running_settings
 from nucliadb_utils.utilities import start_audit_utility, stop_audit_utility
 
 
 async def initialize() -> None:
-    await start_ingest()
+    await start_ingest(SERVICE_NAME)
     await start_audit_utility()
     logging.basicConfig(
         level=logging.INFO,
