+++ nucliadb_search/requirements-sources.txt
@@ -4,3 +4,4 @@
 -e ./nucliadb_ingest
 -e ./nucliadb_swim
 -e ./nucliadb_cluster
+-e ./nucliadb_telemetry
