+++ nucliadb_telemetry/nucliadb_telemetry/settings.py
@@ -5,6 +5,7 @@ class TelemetrySettings(BaseSettings):
     jaeger_host: str = "jaeger.observability.svc.cluster.local"
     jaeger_port: int = 6831
     jeager_enabled: bool = False
+    jaeger_http_port: int = 16686
 
 
 telemetry_settings = TelemetrySettings()
