+++ nucliadb_telemetry/nucliadb_telemetry/tests/conftest.py
@@ -1,4 +1,5 @@
 pytest_plugins = [
+    "pytest_docker_fixtures",
     "nucliadb_utils.tests.nats",
     "nucliadb_telemetry.tests.telemetry",
 ]
