+++ nucliadb_telemetry/nucliadb_telemetry/tests/grpc/test_telemetry.py
@@ -3,7 +3,8 @@ import asyncio
 import pytest
 from httpx import AsyncClient
 
-from nucliadb_telemetry.tests.telemetry import JAEGGER_MESSAGES, Greeter
+from nucliadb_telemetry.tests.telemetry import Greeter
+from nucliadb_telemetry.settings import telemetry_settings
 
 
 @pytest.mark.asyncio
@@ -25,7 +26,18 @@ async def test_telemetry(http_service: AsyncClient, greeter: Greeter):
         == "f13dc5318bf3bef64a0a5ea607db93a1"
     )
 
-    for i in range(10):
-        if len(JAEGGER_MESSAGES) == 0:
+    await asyncio.sleep(2)
+    client = AsyncClient()
+    for _ in range(10):
+        resp = await client.get(
+            f"http://localhost:{telemetry_settings.jaeger_http_port}/api/traces/f13dc5318bf3bef64a0a5ea607db93a1",
+            headers={"Accept": "application/json"},
+        )
+        if resp.status_code != 200 or len(resp.json()["data"][0]["spans"]) < 9:
             await asyncio.sleep(2)
-    assert len(JAEGGER_MESSAGES) > 0
+        else:
+            break
+
+    assert resp.json()["data"][0]["traceID"] == "f13dc5318bf3bef64a0a5ea607db93a1"
+    assert len(resp.json()["data"][0]["spans"]) == 9
+    assert len(resp.json()["data"][0]["processes"]) == 3
