+++ nucliadb_telemetry/nucliadb_telemetry/tests/telemetry.py
@@ -13,67 +13,70 @@ from opentelemetry.propagate import set_global_textmap
 from opentelemetry.propagators.b3 import B3MultiFormat
 
 from nucliadb_telemetry.settings import telemetry_settings
-from nucliadb_telemetry.telemetry import (
-    JetStreamContextTelemetry,
-    OpenTelemetryGRPC,
-    init_telemetry,
-)
+from nucliadb_telemetry.grpc import OpenTelemetryGRPC
+from nucliadb_telemetry.jetstream import JetStreamContextTelemetry
+from nucliadb_telemetry.utils import get_telemetry
 from nucliadb_telemetry.tests.grpc import helloworld_pb2, helloworld_pb2_grpc
+from pytest_docker_fixtures import images  # type: ignore
+from pytest_docker_fixtures.containers._base import BaseImage  # type: ignore
+import requests
+import os
 
-JAEGGER_MESSAGES = []
-
-
-class DefaultProtocol(DatagramProtocol):
-    def datagram_received(self, data, _addr):
-        JAEGGER_MESSAGES.append(data)
-
-
-class UDPServer:
-    __slots__ = ("_port", "_loop", "_protocol", "_transport")
-
-    def __init__(
-        self,
-    ) -> None:
-
-        self._port = _unused_udp_port()
-        self._loop = asyncio.get_event_loop()
-        self._transport = None
-
-    @property
-    def port(self):
-        return self._port
-
-    async def initialize(self):
-        listen = self._loop.create_datagram_endpoint(
-            DefaultProtocol, local_addr=("0.0.0.0", self._port)
-        )
-        self._transport, _ = await listen
-
-    async def finalize(self):
-        self._transport.close()
-        self._transport = None
+images.settings["jaeger"] = {
+    "image": "jaegertracing/all-in-one",
+    "version": "1.33",
+    "options": {
+        "ports": {"6831/udp": None, "16686": None},
+    },
+}
 
 
-def _unused_udp_port() -> int:
-    with contextlib.closing(socket.socket(type=socket.SOCK_DGRAM)) as sock:
-        sock.bind(("127.0.0.1", 0))
-        return sock.getsockname()[1]
+class Jaeger(BaseImage):
+    name = "jaeger"
+    port = 6831
+
+    def get_port(self, port=None):
+        if os.environ.get("TESTING", "") == "jenkins" or "TRAVIS" in os.environ:
+            return port if port else self.port
+        network = self.container_obj.attrs["NetworkSettings"]
+        service_port = "{0}/udp".format(port if port else self.port)
+        for netport in network["Ports"].keys():
+            if netport == service_port:
+                return network["Ports"][service_port][0]["HostPort"]
+
+    def get_http_port(self, port=None):
+        if os.environ.get("TESTING", "") == "jenkins" or "TRAVIS" in os.environ:
+            return 16686
+        network = self.container_obj.attrs["NetworkSettings"]
+        service_port = "16686/tcp"
+        for netport in network["Ports"].keys():
+            if netport == service_port:
+                return network["Ports"][service_port][0]["HostPort"]
+
+    def check(self):
+        resp = requests.get(f"http://{self.host}:{self.get_http_port()}")
+        return resp.status_code == 200
 
 
 @pytest.fixture(scope="function")
-async def settings():
-    server = UDPServer()
-    await server.initialize()
+async def jaeger_server():
+    server = Jaeger()
+    server.run()
+    yield server
+    server.stop()
+
+
+@pytest.fixture(scope="function")
+async def settings(jaeger_server: Jaeger):
     telemetry_settings.jeager_enabled = True
     telemetry_settings.jaeger_host = "127.0.0.1"
-    telemetry_settings.jaeger_port = server.port
-    yield
-    await server.finalize()
+    telemetry_settings.jaeger_port = jaeger_server.get_port()
+    telemetry_settings.jaeger_http_port = jaeger_server.get_http_port()
 
 
 @pytest.fixture(scope="function")
 async def telemetry_grpc(settings):
-    tracer_provider = init_telemetry("HTTP_SERVICE")
+    tracer_provider = get_telemetry("GRPC_SERVICE")
     util = OpenTelemetryGRPC("test_telemetry", tracer_provider)
     yield util
 
@@ -83,10 +86,13 @@ class Greeter(helloworld_pb2_grpc.GreeterServicer):
         self.natsd = natsd
         self.nc = None
         self.subscription = None
+        self.tracer_provider = None
         self.messages = []
 
     async def subscription_worker(self, msg: Msg):
-        self.messages.append(msg)
+        tracer = self.tracer_provider.get_tracer("message_worker")
+        with tracer.start_as_current_span("message_worker_span") as span:
+            self.messages.append(msg)
 
     async def initialize(self):
         self.nc = await nats.connect(servers=[self.natsd])
@@ -98,9 +104,9 @@ class Greeter(helloworld_pb2_grpc.GreeterServicer):
 
         await self.js.add_stream(name="testing", subjects=["testing.*"])
 
-        tracer_provider = init_telemetry("NATS_SERVICE")
+        self.tracer_provider = get_telemetry("NATS_SERVICE")
         self.jsotel = JetStreamContextTelemetry(
-            self.js, "nats_service", tracer_provider
+            self.js, "nats_service", self.tracer_provider
         )
 
         self.subscription = await self.jsotel.subscribe(
@@ -120,7 +126,7 @@ class Greeter(helloworld_pb2_grpc.GreeterServicer):
 
 
 @pytest.fixture(scope="function")
-async def greeter(natsd: str):
+async def greeter(settings, natsd: str):
     obj = Greeter(natsd)
     await obj.initialize()
     yield obj
@@ -139,17 +145,19 @@ async def grpc_service(telemetry_grpc: OpenTelemetryGRPC, greeter: Greeter):
 
 
 @pytest.fixture(scope="function")
-async def http_service(telemetry_grpc: OpenTelemetryGRPC, grpc_service: int):
-    tracer_provider = init_telemetry("HTTP_SERVICE")
+async def http_service(settings, telemetry_grpc: OpenTelemetryGRPC, grpc_service: int):
+    tracer_provider = get_telemetry("HTTP_SERVICE")
     app = FastAPI(title="Test API")  # type: ignore
     set_global_textmap(B3MultiFormat())
     FastAPIInstrumentor.instrument_app(app, tracer_provider=tracer_provider)
 
     @app.get("/")
     async def simple_api():
-        channel = telemetry_grpc.init_client(f"localhost:{grpc_service}")
-        stub = helloworld_pb2_grpc.GreeterStub(channel)
-        response = await stub.SayHello(helloworld_pb2.HelloRequest(name="you"))
+        tracer = tracer_provider.get_tracer(__name__)
+        with tracer.start_as_current_span("simple_api_work") as span:
+            channel = telemetry_grpc.init_client(f"localhost:{grpc_service}")
+            stub = helloworld_pb2_grpc.GreeterStub(channel)
+            response = await stub.SayHello(helloworld_pb2.HelloRequest(name="you"))
         return response.message
 
     client_base_url = "http://test"
