+++ nucliadb_writer/requirements-sources.txt
@@ -4,3 +4,4 @@
 -e ./nucliadb_swim
 -e ./nucliadb_cluster
 -e ./nucliadb_models
+-e ./nucliadb_telemetry
